pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ln2md"
version = "0.1.0"
description = "CLI to convert developer resumes into AI-agent-readable knowledge profiles."
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "ln2md maintainers", email = "maintainers@example.com" }
]
keywords = ["resume", "cli", "ai", "developer-tools", "knowledge-graph"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Utilities",
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pypdf>=4.3.1",
  "pdfplumber>=0.11.4",
  "google-genai>=0.8.0",
  "PyYAML>=6.0.2",
]

[project.scripts]
ln2md = "ln2md.cli:run"

[project.urls]
Homepage = "https://github.com/your-org/ln2md"
Repository = "https://github.com/your-org/ln2md"
Issues = "https://github.com/your-org/ln2md/issues"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["ln2md*"]
